fix: headers lengths

+++ mqtt/src/control_packets/mqtt_connect/variable_header.rs
@@ -14,8 +14,8 @@ pub struct ConnectVariableHeader {
 }
 
 impl ConnectVariableHeader {
-    pub fn length(&self) -> usize {
-        2 + self.protocol_name.length as usize + 1 + 1 + 2 + self.properties.bytes_length
+    pub fn length(&self) -> u8 {
+        2 + self.protocol_name.length as u8 + 1 + 1 + 2 + self.properties.bytes_length
     }
 
     pub fn new(
